GG-7: Add `GET /gadgets` API
In the scope of this app, a *gadget* is a generic device with approximate
characteristics. They are persisted and retrieved, when requested.
The commit introduces an endpoint to get the gadgets in pages. It also
adds a Liquibase migration with a schema for `gadget` and tests.

Other changes:

- The module `domain` is renamed to `persistence` to avoid confusion.
Some domain logic (unit conversion) is planned to be in `service`

- Bumps Spring Boot version to '3.2.4'