GG-14: Add search by name
A user may want to search for specific gadgets. The commit adds an ability
to search them by the name. The new query parameter is introduced:

```http
GET /gadgets?search=word+another
```

The search is case-insensitive and multiple spaces are treated as single.
Both the `%20` and `+` encodings are supported.
For some reason, Spring doesn't decode the query param automatically, so
it is done manually. Also the response API is extended with a new
field, containing the search string.