GG-15: Add `GET /inventory/summary` API
Implements the endpoint to get a summary of gadgets,
currently containing the total volume.

Other changes:

- Updates the package structure

- Removes `updatable = false` and `insertable = false` from the Gadget's primary key.
Initially "insertable" was planned to be used, so that the DB generates
the ID instead of Hibernate. However, the Hibernate validation fails, because the
entity's ID can't be null (although the DB would generate it). Since `GeneratedValue`
has been added, those params are no longer needed

- Minor style adjustments