GG-27: Use Testcontainers
A test database is used in the API / persistence tests.
Previously, the local container with Postgres had a database for `test`.
The commit introduces the Testcontainers DB instead.
It's done, so that the tests don't require the container running constantly on the specific port.
Also it will be easier to add new services in the tests (e.g. MinIO)